Final project added

+++ playlist_generator/src/main/java/track_ninja/playlist_generator/models/Track.java
@@ -31,7 +31,7 @@ public class Track {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = TRACK_ID)
-    private Long trackId;
+    private int trackId;
 
     @Column(name = TITLE)
     private String title;
